requirements.txt
aiohttp
Pillow
numpy